chore(benchmarks): tidy up benchmark stuffs
Delete leftover script files. Add minor change to Jupyter Notebook
(`"ec2_metadata"` key in results table). Simplify pyproject.toml
dependencies list. Change some example parameters in the Hydra config
files. Delete unit test (not meaningful, for now). Rework further the
README. Tune the utils/prepare_nvme.sh to work for both Amazon Linux and
Ubuntu EC2 instances.

+++ s3torchbenchmarking/utils/prepare_nvme.sh
@@ -1,16 +1,20 @@
 #!/usr/bin/env bash
 #
-# Mount an NVMe drive (by default, at `./nvme/`) relative to where this script is run. If a drive is already mounted at
+# Mount an NVMe drive (by default, at `./nvme/`, relative to where this script is run). If a drive is already mounted at
 # the specified location, clear its content.
 
 nvme_dir=${1:-"./nvme/"} # default value
 
 if ! mountpoint -q "$nvme_dir"; then
   rm -rf "$nvme_dir"
-  sudo mkfs -t xfs /dev/nvme1n1
   mkdir -p "$nvme_dir"
-  sudo mount /dev/nvme1n1 "$nvme_dir"
+
+  if grep -q 'NAME="Amazon Linux"' /etc/os-release; then
+    sudo mkfs -t xfs /dev/nvme1n1
+    sudo mount /dev/nvme1n1 "$nvme_dir"
+  elif grep -q 'NAME="Ubuntu"' /etc/os-release; then
+    sudo mount /dev/mapper/vg.01-lv_ephemeral "$nvme_dir"
+  fi
+
   sudo chmod 777 "$nvme_dir"
-else
-  rm -rf "${nvme_dir:?}"/* # https://www.shellcheck.net/wiki/SC2115
 fi
